Added the Routine DataToys to MinimizerFCN that will throw the data according to its errors and produce likelihood distributions of the Data-ThrownData.
 -> file modified:    parameters/config.xml
 -> file modified:    src/FCN/JointFCN.cxx
 -> file modified:    src/FCN/JointFCN.h
 -> file modified:    src/FitBase/JointMeas1D.cxx
 -> file modified:    src/FitBase/JointMeas1D.h
 -> file modified:    src/FitBase/Measurement1D.cxx
 -> file modified:    src/FitBase/Measurement1D.h
 -> file modified:    src/FitBase/Measurement2D.cxx
 -> file modified:    src/FitBase/Measurement2D.h
 -> file modified:    src/FitBase/MeasurementBase.h
 -> file modified:    src/Routines/MinimizerRoutines.cxx
 -> file modified:    src/Routines/MinimizerRoutines.h

Made sure fIsDiag was set to true in the case where no covariance was found during
finalise samplesettings so that time isn't wasted using the fullcovariance (which has already been made diagonal)
to calculate the chi2.

 -> file modified:    src/FitBase/Measurement1D.cxx
 -> file modified:    src/FitBase/JointMeas1D.cxx
 -> file modified:    src/FitBase/Measurement2D.cxx

Fixed a bug in ThrowCovariance where the covariance units were applied also to the data histogram.
Fixed a bug in GetLikelihoodFromEventRate where the function would return nan when the data was
accidentally thrown to negative values.

 -> file modified:    src/Utils/StatUtils.cxx

Changes to MINERvAUtils to fix build procedure. Added preliminary EavQ2 plots to the MINERvA CCQE dataset.

 -> file modified:    src/MINERvA/MINERvAUtils.cxx
 -> file modified:    src/MINERvA/MINERvAUtils.h
 -> file modified:    src/MINERvA/MINERvA_CCQE_XSec_1DQ2_nu.cxx
 -> file modified:    src/MINERvA/MINERvA_CCQE_XSec_1DQ2_nu.h

Added MAXEVENTS setting to GENIEInputHandler.

 -> file modified:   src/InputHandler/GENIEInputHandler.cxx