Added datafiles for MINERvA nue Q2 ratio        new file:   data/MINERvA/CC0pi/MINERvA_CC0pi_nue_Q2Ratio_cov.txt        new file:   data/MINERvA/CC0pi/MINERvA_CC0pi_nue_Q2Ratio_data.txt
Added a throw statement for if Full and Fast reconfigures don't match in the likelihoods.
Requires one extra reconfigure call but is worth it to stop people getting dodgy fits
when using samples that may not be setup for FAST reconfigures.
       modified:   src/FCN/JointFCN.cxx

Updated covariance scaling. It was out by a factor of 10.0
       modified:   src/MINERvA/MINERvA_CC0pi_XSec_1DEe_nue.cxx
       modified:   src/MINERvA/MINERvA_CC0pi_XSec_1DQ2_nue.cxx
       modified:   src/MINERvA/MINERvA_CC0pi_XSec_1DThetae_nue.cxx

Fixed covariance path issues. 1DEnu and 1DEpi covariances are still dodgy
for this dataset though (singular)
       modified:   src/MINERvA/MINERvA_CCCOHPI_XSec_1DEnu_nu.cxx
       modified:   src/MINERvA/MINERvA_CCCOHPI_XSec_1DEpi_antinu.cxx

Added CC1pip 2017 signal selection and fixed Tpi bug in CC1pip 2017 data
       modified:   src/MINERvA/MINERvA_CC1pip_XSec_1D_2017Update.cxx
       modified:   src/MINERvA/MINERvA_SignalDef.cxx
       modified:   src/MINERvA/MINERvA_SignalDef.h