Added MINERvA CC1pi0 numu 2017
Added signal definition function to namespace. Although not strictly
necessary (just calls isCC1pi) I chose to keep it in because we could
put the W cut in this function.

Added function for CosThetaAdler and PhiAdler to FitUtils namespace

The Wcut is implemented in the experiment class to allow us to look at
events that don't pass the W cut.