Adds a simple minimizer test and Tweaks some debugging in the core fitter, but no significant changes.
- Minimizer tests creates 3 dummy events, each of different
interaction modes and uses them to build some weighted fake data
the minimizer is then given the normalisation dials for these
interaction modes and tries to recover the fake data weights
from a histogram binned in true modes -- i.e. has the exact
orthogonal freedom to get the answer right. This is a test
of the interface to the fitter and the error estimation in
a completely idealised scenario.